requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
httpx
pyjwt
pydantic
